+++ R/geom_ideogram.R
@@ -39,22 +39,31 @@ geom_ideogram <- function(genome = "hg19", mark.color = "red", mark.alpha = 0.7,
 
 #' @export
 ggplot_add.ideogram <- function(object, plot, object_name) {
-  if (length(plot$layers) == 0) {
-    # geom_base
-    # get plot data
+  # if (length(plot$layers) == 0) {
+  #   # geom_base
+  #   # get plot data
+  #   plot.data <- plot[[1]]$layers[[1]]$data
+  #   # prepare plot range
+  #   plot.chr <- as.character(plot.data[1, "seqnames"])
+  #   plot.region.start <- plot.data[1, "start"]
+  #   plot.region.end <- plot.data[nrow(plot.data), "end"]
+  # } else {
+  #   # get plot data
+  #   plot.data <- plot$layers[[1]]$data
+  #   # prepare plot range
+  #   plot.chr <- as.character(plot.data[1, "seqnames"])
+  #   plot.region.start <- plot$coordinates$limits$x[1]
+  #   plot.region.end <- plot$coordinates$limits$x[2]
+  # }
+  # get plot data, plot data should contain bins
+  if ("patchwork" %in% class(plot)) {
     plot.data <- plot[[1]]$layers[[1]]$data
-    # prepare plot range
-    plot.chr <- as.character(plot.data[1, "seqnames"])
-    plot.region.start <- plot.data[1, "start"]
-    plot.region.end <- plot.data[nrow(plot.data), "end"]
   } else {
-    # get plot data
     plot.data <- plot$layers[[1]]$data
-    # prepare plot range
-    plot.chr <- as.character(plot.data[1, "seqnames"])
-    plot.region.start <- plot$coordinates$limits$x[1]
-    plot.region.end <- plot$coordinates$limits$x[2]
   }
+  plot.chr <- as.character(plot.data[1, "seqnames"])
+  plot.region.start <- plot.data[1, "start"]
+  plot.region.end <- plot.data[nrow(plot.data), "end"]
 
   # get parameters
   genome <- object$genome