+++ R/ggcoverage.R
@@ -82,8 +82,8 @@ ggcoverage <- function(data, region = "chr14:21,677,306-21,737,601", gtf.gr = NU
   # create plot
   coverage.plot <- ggplot() +
     geom_coverage(
-      data = data, mapping = mapping,
-      color = color, rect.color = rect.color, facet.key = facet.key, facet.order = facet.color, facet.color = facet.color,
+      data = data, mapping = mapping, color = color, rect.color = rect.color,
+      single.nuc = single.nuc, facet.key = facet.key, facet.order = facet.color, facet.color = facet.color,
       group.key = group.key, range.size = range.size, range.position = range.position,
       mark.region = mark.region, mark.color = mark.color, mark.alpha = mark.alpha, show.mark.label = show.mark.label,
       mark.label.size = mark.label.size
@@ -91,9 +91,9 @@ ggcoverage <- function(data, region = "chr14:21,677,306-21,737,601", gtf.gr = NU
   # add theme
   if (range.position == "in") {
     coverage.plot +
-      theme_coverage(space = plot.space, x.range = c(plot.range.start, plot.range.end))
+      theme_coverage(space = plot.space)
   } else if (range.position == "out") {
     coverage.plot +
-      theme_coverage2(space = plot.space, x.range = c(plot.range.start, plot.range.end))
+      theme_coverage2(space = plot.space)
   }
 }